% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minigpkg-package.R
\docType{package}
\name{minigpkg-package}
\alias{minigpkg}
\alias{minigpkg-package}
\title{minigpkg: Experimental 'GeoPackage' Input and Output}
\description{
Implements experimental 'GeoPackage' input and output.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0001-8049-7069}{ORCID})

}
\keyword{internal}
